fix(agent): the endpoint offered the wrong port when the operator changed nothing
`RankAssignment::settings` is a **sparse diff** — the operator's overrides,
not the effective settings. The endpoint read it as though it were the whole
truth, so a recipe pinning `port: 8888` and an operator who overrode nothing
produced an empty map and a URL reading `:8000`, while the model served on
`:8888`. That URL is the one thing on the screen an operator actually pastes
somewhere.

The port now resolves in three layers: what the operator set, then what the
recipe pins, then the serving runtime's default. The middle layer was missing
entirely. It is asked of the machine through a new `RankService::recipe_port`
rather than assumed, for the same reason the content hash is: the answer lives
in that machine's own copy of the recipe. `None` is not an error there — it
means the recipe pins no port and the runtime's default applies, a number this
agent does not know.

The fixture rank constructors now build by difference from `ready`, so the
next field added to the service does not have to be restated in four places to
compile.

cases.rs crossed the 500-line cap, so the teardown orderings move to
clusterdriver/teardown.rs by exact piecewise copy — a deliberate stop, a rank
that dies on startup, and a rank that dies after the operator was told it was
up. 209 tests pass, the same set as before the split.

2 new tests, verified to fail without this change:
  - the endpoint uses the recipe's port when nothing was overridden
  - an override still beats the recipe, or the setting would do nothing
